+++ lib/onlyoffice_digitalocean_wrapper/digitalocean_wrapper/logger_wrapper.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require 'logger'
+
+module OnlyofficeDigitaloceanWrapper
+  # Logger module for logging stuff
+  module LoggerWrapper
+    # @return [Logger] default logger
+    def logger
+      @logger ||= Logger.new($stdout)
+    end
+  end
+end